chore: project audit — ship 15 fixes across security / perf / refactor
Security:
- Expiring + localOnly pasteboard for API keys (AdminApiKeysScreen)
- Redact api_key= / token query items from debug URL logs (CinemaxKit
  + VideoPlayerView)

Performance:
- Rainbow accent animation 30Hz → 10Hz; gated on motionEffects; hooked
  motionEffectsDidChange() from AppNavigation
- ImagePipeline.shared config idempotent via static flag (survives
  SwiftUI root struct recreation)
- New ImageURLBuilder.backdropPixelWidth = min(screenPixelWidth, 1920)
  used by hero backdrops (HomeScreen / MediaDetailScreen / LibraryHero)
- Clear APICache on user switch to prevent cross-user DTO bleed

Refactor:
- Split JellyfinAPIClient 1098 → 218 (core) + Admin / Library / Playback
  extensions
- Extract ChapterController + EndOfSeriesOverlayController from
  NativeVideoPresenter (1256 → 979), following the existing
  PlaybackReporter / SkipSegmentController / SleepTimerController pattern
- Replace 3 raw UserDefaults string keys in NativeVideoPresenter with
  SettingsKey.*
- Route CastCircle and tvOS user avatar through CinemaLazyImage (was
  raw LazyImage / AsyncImage)
- Delete unreferenced AdminComingSoonScreen (0 call sites after P3c)

Quality:
- Surface try? failures at 3 user-initiated sites (clearContinueWatching,
  device list load, tvOS user list fallback) with 4 new localized toasts
  (en/fr)
- CinemaLazyImage bare .font(.largeTitle) → CinemaScale.pt(28)
- HomeScreen "LIVE" pill: hardcoded size 10 → CinemaScale.pt(10) +
  localization key home.liveSession
- MediaDetailScreen season-pill hardcoded padding → CinemaSpacing.spacing3

Tests:
- 4 new PlaybackReporter behaviour tests (nil-context no-op, reset,
  start)

Docs:
- Add AUDIT.md (full-project audit report)
- Compress CLAUDE.md 49k → 38k chars (under perf threshold)
- Annotate project.yml CinemaxTests target explaining the
  Get-double-link constraint that keeps tests in Tests/

SwiftUI previews:
- Add #Preview for CinemaButton, PosterCard, EmptyStateView,
  CinemaLazyImage, RatingBadge, CinemaToggleIndicator, CastCircle,
  ErrorStateView

All builds green (iOS + tvOS). 40/40 tests pass.
