+++ Packages/CinemaxKit/Sources/CinemaxKit/Networking/JellyfinAPIClient.swift
@@ -13,6 +13,28 @@ func debugLog(_ message: String) {
 }
 #endif
 
+/// Strips secret query items (`api_key`, `ApiKey`, `X-Emby-Token`, anything
+/// containing "token") from a URL so logs can include the URL without leaking
+/// the access token. Keeps path + non-secret query items for debuggability.
+/// Accepts `String?` because Jellyfin's `transcodingURL` is a string path we
+/// log before resolving to a `URL`.
+public func redactedURL(_ raw: String?) -> String {
+    guard let raw, !raw.isEmpty else { return "nil" }
+    guard var components = URLComponents(string: raw) else { return raw }
+    if let items = components.queryItems {
+        components.queryItems = items.map { item in
+            let lower = item.name.lowercased()
+            if lower == "api_key" || lower == "apikey" || lower.contains("token") {
+                return URLQueryItem(name: item.name, value: "REDACTED")
+            }
+            return item
+        }
+    }
+    return components.string ?? raw
+}
+
+public func redactedURL(_ url: URL) -> String { redactedURL(url.absoluteString) }
+
 public final class JellyfinAPIClient: Sendable {
     // `JellyfinClient` (from jellyfin-sdk-swift) is not marked `Sendable`, but we need
     // cross-actor access. Invariant: every access to `_jellyfinClient` / `_serverURL`
@@ -22,30 +44,30 @@ public final class JellyfinAPIClient: Sendable {
     nonisolated(unsafe) private var _jellyfinClient: JellyfinClient?
     nonisolated(unsafe) private var _serverURL: URL?
     nonisolated(unsafe) private var _maxContentAge: Int = 0
-    private let cache = APICache()
+    internal let cache = APICache()
 
     public init() {}
 
-    private func getClient() -> JellyfinClient? {
+    internal func getClient() -> JellyfinClient? {
         lock.lock()
         defer { lock.unlock() }
         return _jellyfinClient
     }
 
-    private func setClient(_ client: JellyfinClient, url: URL) {
+    internal func setClient(_ client: JellyfinClient, url: URL) {
         lock.lock()
         defer { lock.unlock() }
         _jellyfinClient = client
         _serverURL = url
     }
 
-    private func getServerURL() -> URL? {
+    internal func getServerURL() -> URL? {
         lock.lock()
         defer { lock.unlock() }
         return _serverURL
     }
 
-    private func getMaxContentAge() -> Int {
+    internal func getMaxContentAge() -> Int {
         lock.lock()
         defer { lock.unlock() }
         return _maxContentAge
@@ -63,7 +85,7 @@ public final class JellyfinAPIClient: Sendable {
 
     /// Keeps only items whose `officialRating` passes the active maximum-age
     /// limit. A no-op when the limit is disabled.
-    private func applyRatingFilter(_ items: [BaseItemDto]) -> [BaseItemDto] {
+    internal func applyRatingFilter(_ items: [BaseItemDto]) -> [BaseItemDto] {
         let maxAge = getMaxContentAge()
         guard maxAge > 0 else { return items }
         return items.filter { ContentRatingClassifier.passes(rating: $0.officialRating, maxAge: maxAge) }
@@ -154,907 +176,6 @@ public final class JellyfinAPIClient: Sendable {
         )
     }
 
-    // MARK: - Users
-
-    public func getPublicUsers() async throws -> [UserDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getPublicUsers)
-        return response.value
-    }
-
-    public func getUsers() async throws -> [UserDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getUsers())
-        return response.value
-    }
-
-    /// Returns all active sessions on the server. Used by the "Currently watching" indicator
-    /// on Home to show what other users are streaming right now.
-    public func getActiveSessions(activeWithinSeconds: Int = 60) async throws -> [SessionInfoDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetSessionsParameters(activeWithinSeconds: activeWithinSeconds)
-        let response = try await client.send(Paths.getSessions(parameters: params))
-        return response.value
-    }
-
-    // MARK: - Media Queries
-
-    public func getResumeItems(userId: String, limit: Int = 10) async throws -> [BaseItemDto] {
-        let cacheKey = "resume-\(userId)-\(limit)-\(getMaxContentAge())"
-        if let cached: [BaseItemDto] = cache.get(cacheKey) { return cached }
-
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetResumeItemsParameters(
-            userID: userId,
-            limit: limit,
-            enableUserData: true,
-            enableImageTypes: [.primary, .backdrop, .thumb]
-        )
-        let response = try await client.send(Paths.getResumeItems(parameters: params))
-        let result = applyRatingFilter(response.value.items ?? [])
-        cache.set(cacheKey, value: result, ttl: 30)
-        return result
-    }
-
-    public func getLatestMedia(userId: String, parentId: String? = nil, limit: Int = 16) async throws -> [BaseItemDto] {
-        let cacheKey = "latest-\(userId)-\(limit)-\(getMaxContentAge())"
-        if let cached: [BaseItemDto] = cache.get(cacheKey) { return cached }
-
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetLatestMediaParameters(
-            userID: userId,
-            parentID: parentId,
-            enableImages: true,
-            imageTypeLimit: 1,
-            enableUserData: true,
-            limit: limit
-        )
-        let response = try await client.send(Paths.getLatestMedia(parameters: params))
-        let result = applyRatingFilter(response.value)
-        cache.set(cacheKey, value: result, ttl: 60)
-        return result
-    }
-
-    public func getItems(
-        userId: String,
-        parentId: String? = nil,
-        includeItemTypes: [BaseItemKind]? = nil,
-        sortBy: [ItemSortBy]? = nil,
-        sortOrder: [JellyfinAPI.SortOrder]? = nil,
-        genres: [String]? = nil,
-        years: [Int]? = nil,
-        isFavorite: Bool? = nil,
-        filters: [ItemFilter]? = nil,
-        limit: Int? = nil,
-        startIndex: Int? = nil
-    ) async throws -> (items: [BaseItemDto], totalCount: Int) {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let maxOfficialRating = ContentRatingClassifier.maxOfficialRatingCode(forAge: getMaxContentAge())
-        let params = Paths.GetItemsParameters(
-            userID: userId,
-            maxOfficialRating: maxOfficialRating,
-            startIndex: startIndex,
-            limit: limit,
-            isRecursive: true,
-            sortOrder: sortOrder,
-            parentID: parentId,
-            includeItemTypes: includeItemTypes,
-            filters: filters,
-            sortBy: sortBy,
-            genres: genres,
-            years: years
-        )
-        let response = try await client.send(Paths.getItems(parameters: params))
-        let result = response.value
-        return (result.items ?? [], result.totalRecordCount ?? 0)
-    }
-
-    /// Fetches available genres for the given item types from the server.
-    public func getGenres(
-        userId: String,
-        includeItemTypes: [BaseItemKind]? = nil
-    ) async throws -> [String] {
-        let itemTypes = includeItemTypes?.map(\.rawValue).sorted().joined(separator: ",") ?? ""
-        let cacheKey = "genres-\(userId)-\(itemTypes)"
-        if let cached: [String] = cache.get(cacheKey) { return cached }
-
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetQueryFiltersParameters(
-            userID: userId,
-            includeItemTypes: includeItemTypes,
-            isRecursive: true
-        )
-        let response = try await client.send(Paths.getQueryFilters(parameters: params))
-        let result = response.value.genres?.compactMap(\.name) ?? []
-        cache.set(cacheKey, value: result, ttl: 300)
-        return result
-    }
-
-    public func getUserViews(userId: String) async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetUserViewsParameters(userID: userId)
-        let response = try await client.send(Paths.getUserViews(parameters: params))
-        return response.value.items ?? []
-    }
-
-    public func getItem(userId: String, itemId: String) async throws -> BaseItemDto {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getItem(itemID: itemId, userID: userId))
-        return response.value
-    }
-
-    public func getSimilarItems(itemId: String, userId: String, limit: Int = 12) async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetSimilarItemsParameters(userID: userId, limit: limit)
-        let response = try await client.send(Paths.getSimilarItems(itemID: itemId, parameters: params))
-        return applyRatingFilter(response.value.items ?? [])
-    }
-
-    public func searchItems(userId: String, searchTerm: String, limit: Int = 20) async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let maxOfficialRating = ContentRatingClassifier.maxOfficialRatingCode(forAge: getMaxContentAge())
-        let params = Paths.GetItemsParameters(
-            userID: userId,
-            maxOfficialRating: maxOfficialRating,
-            limit: limit,
-            isRecursive: true,
-            searchTerm: searchTerm,
-            includeItemTypes: [.movie, .series, .episode]
-        )
-        let response = try await client.send(Paths.getItems(parameters: params))
-        return response.value.items ?? []
-    }
-
-    public func getSeasons(seriesId: String, userId: String) async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetSeasonsParameters(userID: userId, enableUserData: true)
-        let response = try await client.send(Paths.getSeasons(seriesID: seriesId, parameters: params))
-        return response.value.items ?? []
-    }
-
-    public func getEpisodes(seriesId: String, seasonId: String, userId: String) async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetEpisodesParameters(userID: userId, fields: [.overview], seasonID: seasonId, enableUserData: true)
-        let response = try await client.send(Paths.getEpisodes(seriesID: seriesId, parameters: params))
-        return applyRatingFilter(response.value.items ?? [])
-    }
-
-    public func getNextUp(seriesId: String, userId: String) async throws -> BaseItemDto? {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetNextUpParameters(
-            userID: userId,
-            limit: 1,
-            seriesID: seriesId,
-            enableUserData: true
-        )
-        let response = try await client.send(Paths.getNextUp(parameters: params))
-        guard let next = response.value.items?.first else { return nil }
-        return ContentRatingClassifier.passes(rating: next.officialRating, maxAge: getMaxContentAge()) ? next : nil
-    }
-
-    // MARK: - User Item Data
-
-    public func markItemUnplayed(itemId: String, userId: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.markUnplayedItem(itemID: itemId, userID: userId))
-        cache.clear()
-    }
-
-    // MARK: - Devices
-
-    public func getDevices() async throws -> [DeviceInfoDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getDevices())
-        return response.value.items ?? []
-    }
-
-    public func deleteDevice(id: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.deleteDevice(id: id))
-    }
-
-    // MARK: - Admin
-
-    public func getUserByID(id: String) async throws -> UserDto {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getUserByID(userID: id))
-        return response.value
-    }
-
-    public func createUserByName(name: String, password: String?) async throws -> UserDto {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let body = CreateUserByName(name: name, password: password)
-        let response = try await client.send(Paths.createUserByName(body))
-        return response.value
-    }
-
-    public func updateUser(id: String, user: UserDto) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.updateUser(userID: id, user))
-    }
-
-    public func updateUserPolicy(id: String, policy: UserPolicy) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.updateUserPolicy(userID: id, policy))
-    }
-
-    public func updateUserPassword(id: String, newPassword: String, resetPassword: Bool) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        // Admin-as-another-user password resets don't require the target user's
-        // current password — the server trusts the admin's session, so we send
-        // only `newPw` (and `resetPassword` to clear without setting a new one).
-        let body = UpdateUserPassword(
-            currentPassword: nil,
-            currentPw: nil,
-            newPw: resetPassword ? nil : newPassword,
-            isResetPassword: resetPassword
-        )
-        _ = try await client.send(Paths.updateUserPassword(userID: id, body))
-    }
-
-    public func deleteUser(id: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.deleteUser(userID: id))
-    }
-
-    public func getMediaFolders() async throws -> [BaseItemDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getMediaFolders())
-        return response.value.items ?? []
-    }
-
-    public func getActivityLogEntries(startIndex: Int, limit: Int, minDate: Date?) async throws -> (entries: [ActivityLogEntry], total: Int) {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.GetLogEntriesParameters(startIndex: startIndex, limit: limit, minDate: minDate)
-        let response = try await client.send(Paths.getLogEntries(parameters: params))
-        return (response.value.items ?? [], response.value.totalRecordCount ?? 0)
-    }
-
-    public func getSystemInfo() async throws -> SystemInfo {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getSystemInfo)
-        return response.value
-    }
-
-    // MARK: - Plugins
-
-    public func getInstalledPlugins() async throws -> [PluginInfo] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getPlugins)
-        return response.value
-    }
-
-    public func enablePlugin(id: String, version: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.enablePlugin(pluginID: id, version: version))
-    }
-
-    public func disablePlugin(id: String, version: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.disablePlugin(pluginID: id, version: version))
-    }
-
-    public func uninstallPlugin(id: String, version: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.uninstallPluginByVersion(pluginID: id, version: version))
-    }
-
-    // MARK: - Plugin catalog
-
-    public func getPluginCatalog() async throws -> [PackageInfo] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getPackages)
-        return response.value
-    }
-
-    public func installPackage(name: String, assemblyGuid: String?, version: String?, repositoryURL: String?) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.InstallPackageParameters(
-            assemblyGuid: assemblyGuid,
-            version: version,
-            repositoryURL: repositoryURL
-        )
-        _ = try await client.send(Paths.installPackage(name: name, parameters: params))
-    }
-
-    // MARK: - Scheduled tasks
-
-    public func getScheduledTasks(includeHidden: Bool) async throws -> [TaskInfo] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        // `isHidden == false` requests visible-only; `nil` returns everything.
-        let response = try await client.send(Paths.getTasks(isHidden: includeHidden ? nil : false))
-        return response.value
-    }
-
-    public func startTask(id: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.startTask(taskID: id))
-    }
-
-    public func stopTask(id: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.stopTask(taskID: id))
-    }
-
-    public func updateTaskTriggers(id: String, triggers: [TaskTriggerInfo]) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.updateTask(taskID: id, triggers))
-    }
-
-    // MARK: - Encoding options (named configuration)
-
-    public func getEncodingOptions() async throws -> EncodingOptions {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getNamedConfiguration(key: "encoding"))
-        return try JSONDecoder().decode(EncodingOptions.self, from: response.value)
-    }
-
-    public func updateEncodingOptions(_ options: EncodingOptions) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        // Round-trip through AnyJSON — SDK's named-configuration endpoint is
-        // type-erased so we can't hand it an EncodingOptions directly.
-        let data = try JSONEncoder().encode(options)
-        let anyJSON = try JSONDecoder().decode(AnyJSON.self, from: data)
-        _ = try await client.send(Paths.updateNamedConfiguration(key: "encoding", anyJSON))
-    }
-
-    // MARK: - Network configuration
-
-    public func getNetworkConfiguration() async throws -> NetworkConfiguration {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getNamedConfiguration(key: "network"))
-        return try JSONDecoder().decode(NetworkConfiguration.self, from: response.value)
-    }
-
-    public func updateNetworkConfiguration(_ config: NetworkConfiguration) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let data = try JSONEncoder().encode(config)
-        let anyJSON = try JSONDecoder().decode(AnyJSON.self, from: data)
-        _ = try await client.send(Paths.updateNamedConfiguration(key: "network", anyJSON))
-    }
-
-    // MARK: - Logs
-
-    public func getServerLogs() async throws -> [LogFile] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getServerLogs)
-        return response.value
-    }
-
-    public func getLogFileContents(name: String) async throws -> String {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getLogFile(name: name))
-        return response.value
-    }
-
-    // MARK: - API keys
-    //
-    // We intentionally do not cache key responses (unlike resume/latest) — the
-    // list is small, changes at admin speed, and we don't want stale revoked
-    // keys lingering in memory longer than necessary.
-
-    public func getApiKeys() async throws -> [AuthenticationInfo] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getKeys)
-        return response.value.items ?? []
-    }
-
-    public func createApiKey(app: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.createKey(app: app))
-    }
-
-    public func revokeApiKey(key: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.revokeKey(key: key))
-    }
-
-    // MARK: - Metadata editor
-
-    public func updateItem(id: String, item: BaseItemDto) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.updateItem(itemID: id, item))
-        // Invalidate cached item queries — stale data after a metadata edit
-        // is misleading (renamed titles staying renamed only in memory, etc.).
-        cache.clear()
-    }
-
-    public func refreshItem(
-        id: String,
-        metadataMode: MetadataRefreshMode,
-        imageMode: MetadataRefreshMode,
-        replaceAllMetadata: Bool,
-        replaceAllImages: Bool
-    ) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.RefreshItemParameters(
-            metadataRefreshMode: metadataMode,
-            imageRefreshMode: imageMode,
-            isReplaceAllMetadata: replaceAllMetadata,
-            isReplaceAllImages: replaceAllImages
-        )
-        _ = try await client.send(Paths.refreshItem(itemID: id, parameters: params))
-        cache.clear()
-    }
-
-    public func deleteItem(id: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.deleteItem(itemID: id))
-        cache.clear()
-    }
-
-    public func downloadRemoteImage(itemId: String, type: JellyfinAPI.ImageType, imageURL: String) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let params = Paths.DownloadRemoteImageParameters(type: type, imageURL: imageURL)
-        _ = try await client.send(Paths.downloadRemoteImage(itemID: itemId, parameters: params))
-        cache.clear()
-    }
-
-    public func deleteItemImage(id: String, type: JellyfinAPI.ImageType, index: Int?) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        // SDK path takes imageType as String (the wire format is the raw value).
-        if let index {
-            _ = try await client.send(Paths.deleteItemImageByIndex(itemID: id, imageType: type.rawValue, imageIndex: index))
-        } else {
-            _ = try await client.send(Paths.deleteItemImage(itemID: id, imageType: type.rawValue))
-        }
-        cache.clear()
-    }
-
-    public func searchRemoteMovies(query: MovieInfoRemoteSearchQuery) async throws -> [RemoteSearchResult] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getMovieRemoteSearchResults(query))
-        return response.value
-    }
-
-    public func searchRemoteSeries(query: SeriesInfoRemoteSearchQuery) async throws -> [RemoteSearchResult] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getSeriesRemoteSearchResults(query))
-        return response.value
-    }
-
-    public func applyRemoteSearchResult(itemId: String, result: RemoteSearchResult, replaceAllImages: Bool) async throws {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        _ = try await client.send(Paths.applySearchCriteria(itemID: itemId, isReplaceAllImages: replaceAllImages, result))
-        cache.clear()
-    }
-
-    // MARK: - Media Segments
-
-    public func getMediaSegments(itemId: String, includeSegmentTypes: [JellyfinAPI.MediaSegmentType]? = nil) async throws -> [MediaSegmentDto] {
-        guard let client = getClient() else { throw JellyfinError.notConnected }
-        let response = try await client.send(Paths.getItemSegments(itemID: itemId, includeSegmentTypes: includeSegmentTypes))
-        return response.value.items ?? []
-    }
-
-    // MARK: - Playback
-
-    /// Builds the best streaming URL for the given item.
-    /// Follows Swiftfin's exact flow:
-    /// 1. Get full item → extract initial media source
-    /// 2. POST PlaybackInfo with device profile
-    /// 3. Build stream URL from response (transcodingURL or direct stream)
-    public func getPlaybackInfo(itemId: String, userId: String, maxBitrate: Int = 40_000_000, audioStreamIndex: Int? = nil, subtitleStreamIndex: Int? = nil) async throws -> PlaybackInfo {
-        guard let client = getClient(),
-              let serverURL = getServerURL() else {
-            throw JellyfinError.notConnected
-        }
-
-        let token = client.configuration.accessToken
-
-        // Step 1: Get full item details (same as Swiftfin's getFullItem)
-        var item = try await getItem(userId: userId, itemId: itemId)
-        var effectiveItemId = itemId
-
-        #if DEBUG
-        debugLog("Item '\(item.name ?? "?")': type=\(item.type?.rawValue ?? "nil"), mediaSources=\(item.mediaSources?.count ?? 0)")
-        #endif
-
-        // If this is a Series or Season, resolve to a playable episode
-        if item.type == .series {
-            let resolvedId: String
-            if let nextEp = try await getNextUp(seriesId: itemId, userId: userId),
-               let epId = nextEp.id {
-                resolvedId = epId
-                #if DEBUG
-                debugLog("Resolved Series → next up episode (\(epId))")
-                #endif
-            } else {
-                // No next up — try first episode of first season
-                let seasons = try await getSeasons(seriesId: itemId, userId: userId)
-                guard let firstSeason = seasons.first, let seasonId = firstSeason.id else {
-                    throw JellyfinError.playbackFailed("No episodes available")
-                }
-                let episodes = try await getEpisodes(seriesId: itemId, seasonId: seasonId, userId: userId)
-                guard let firstEp = episodes.first, let firstEpId = firstEp.id else {
-                    throw JellyfinError.playbackFailed("No episodes available")
-                }
-                resolvedId = firstEpId
-                #if DEBUG
-                debugLog("Resolved Series → first episode (\(firstEpId))")
-                #endif
-            }
-            item = try await getItem(userId: userId, itemId: resolvedId)
-            effectiveItemId = resolvedId
-        } else if item.type == .season {
-            guard let seriesId = item.seriesID else {
-                throw JellyfinError.playbackFailed("No series ID for season")
-            }
-            let episodes = try await getEpisodes(seriesId: seriesId, seasonId: itemId, userId: userId)
-            guard let firstEp = episodes.first, let firstEpId = firstEp.id else {
-                throw JellyfinError.playbackFailed("No episodes in this season")
-            }
-            item = try await getItem(userId: userId, itemId: firstEpId)
-            effectiveItemId = firstEpId
-            #if DEBUG
-            debugLog("Resolved Season → first episode (\(firstEpId))")
-            #endif
-        }
-
-        let initialMediaSource = item.mediaSources?.first
-        let initialMediaSourceId = initialMediaSource?.id
-
-        #if DEBUG
-        debugLog("Playback item '\(item.name ?? "?")': type=\(item.type?.rawValue ?? "nil"), mediaSources=\(item.mediaSources?.count ?? 0)")
-        if let src = initialMediaSource {
-            debugLog("  source: id=\(src.id ?? "nil"), container=\(src.container ?? "nil")")
-            for s in src.mediaStreams ?? [] {
-                debugLog("  stream: type=\(s.type?.rawValue ?? "?"), codec=\(s.codec ?? "?")")
-            }
-        } else {
-            debugLog("  No media sources from getItem")
-        }
-        #endif
-
-        // Step 2: POST PlaybackInfo (matching Swiftfin exactly)
-        var body = PlaybackInfoDto(deviceProfile: Self.buildAppleDeviceProfile(maxBitrate: maxBitrate))
-        body.isAutoOpenLiveStream = true
-        body.maxStreamingBitrate = maxBitrate
-        body.userID = userId
-        body.audioStreamIndex = audioStreamIndex
-        body.subtitleStreamIndex = subtitleStreamIndex
-        if let initialMediaSourceId {
-            body.mediaSourceID = initialMediaSourceId
-        }
-
-        #if DEBUG
-        debugLog("POST /Items/\(effectiveItemId)/PlaybackInfo (mediaSourceId=\(initialMediaSourceId ?? "nil"))")
-        #endif
-
-        // Use raw URLSession POST to capture the full 500 error body for diagnosis
-        let response: PlaybackInfoResponse
-        do {
-            response = try await rawPostPlaybackInfo(
-                serverURL: serverURL,
-                itemId: effectiveItemId,
-                client: client,
-                body: body
-            )
-        } catch {
-            #if DEBUG
-            debugLog("POST PlaybackInfo failed: \(error) — using direct stream fallback")
-            #endif
-            return buildDirectStreamURL(
-                itemId: effectiveItemId, serverURL: serverURL, token: token, etag: item.etag
-            )
-        }
-
-        // Step 3: Find matching media source in response (same matching as Swiftfin)
-        let mediaSource: MediaSourceInfo? = {
-            guard let sources = response.mediaSources else { return nil }
-            if let etag = initialMediaSource?.eTag,
-               let match = sources.first(where: { $0.eTag == etag }) { return match }
-            if let id = initialMediaSourceId,
-               let match = sources.first(where: { $0.id == id }) { return match }
-            return sources.first
-        }()
-
-        guard let mediaSource else {
-            #if DEBUG
-            debugLog("No media source in PlaybackInfo response — using direct stream fallback")
-            #endif
-            return buildDirectStreamURL(
-                itemId: effectiveItemId, serverURL: serverURL, token: token, etag: item.etag
-            )
-        }
-
-        let playSessionId = response.playSessionID
-
-        #if DEBUG
-        debugLog("PlaybackInfo OK: playSession=\(playSessionId ?? "nil")")
-        debugLog("  source: id=\(mediaSource.id ?? "nil"), container=\(mediaSource.container ?? "nil")")
-        debugLog("  directPlay=\(mediaSource.isSupportsDirectPlay ?? false), directStream=\(mediaSource.isSupportsDirectStream ?? false)")
-        debugLog("  transcodingURL=\(mediaSource.transcodingURL ?? "nil")")
-        #endif
-
-        // Step 4: Build stream URL (same logic as Swiftfin's streamURL)
-
-        // Option A: Server returned a transcoding URL — use it directly
-        if let transcodingPath = mediaSource.transcodingURL {
-            let baseURL = serverURL.absoluteString.trimmingCharacters(in: ["/"])
-            if let url = URL(string: baseURL + transcodingPath) {
-                #if DEBUG
-                debugLog("Transcode URL: \(url)")
-                #endif
-                let (audioTracks, subtitleTracks) = Self.extractTracks(from: mediaSource)
-                return PlaybackInfo(
-                    url: url,
-                    playSessionId: playSessionId,
-                    mediaSourceId: mediaSource.id,
-                    playMethod: .transcode,
-                    audioTracks: audioTracks,
-                    subtitleTracks: subtitleTracks,
-                    selectedAudioIndex: audioStreamIndex ?? mediaSource.defaultAudioStreamIndex,
-                    selectedSubtitleIndex: subtitleStreamIndex ?? mediaSource.defaultSubtitleStreamIndex,
-                    authToken: nil // token already embedded in Jellyfin's HLS URL
-                )
-            }
-        }
-
-        // Option B: Direct stream (Swiftfin uses Paths.getVideoStream with isStatic=true)
-        guard let playSessionId else {
-            return buildDirectStreamURL(
-                itemId: effectiveItemId, serverURL: serverURL, token: token, etag: item.etag
-            )
-        }
-
-        guard var components = URLComponents(url: serverURL, resolvingAgainstBaseURL: false) else {
-            return buildDirectStreamURL(
-                itemId: effectiveItemId, serverURL: serverURL, token: token, etag: item.etag
-            )
-        }
-        components.path = "/Videos/\(effectiveItemId)/stream"
-        var queryItems = [
-            URLQueryItem(name: "static", value: "true"),
-            URLQueryItem(name: "playSessionId", value: playSessionId),
-            URLQueryItem(name: "mediaSourceId", value: mediaSource.id ?? effectiveItemId),
-        ]
-        if let tag = item.etag { queryItems.append(URLQueryItem(name: "tag", value: tag)) }
-        // Token passed via Authorization header on AVURLAsset, not in URL query params
-        components.queryItems = queryItems
-
-        guard let url = components.url else {
-            return buildDirectStreamURL(
-                itemId: effectiveItemId, serverURL: serverURL, token: token, etag: item.etag
-            )
-        }
-
-        #if DEBUG
-        debugLog("Direct stream URL: \(url)")
-        #endif
-        let (audioTracks, subtitleTracks) = Self.extractTracks(from: mediaSource)
-        return PlaybackInfo(
-            url: url,
-            playSessionId: playSessionId,
-            mediaSourceId: mediaSource.id,
-            playMethod: .directStream,
-            audioTracks: audioTracks,
-            subtitleTracks: subtitleTracks,
-            selectedAudioIndex: audioStreamIndex ?? mediaSource.defaultAudioStreamIndex,
-            selectedSubtitleIndex: subtitleStreamIndex ?? mediaSource.defaultSubtitleStreamIndex,
-            authToken: token
-        )
-    }
-
-    /// Raw HTTP POST to PlaybackInfo, captures the full response body for diagnosis.
-    private func rawPostPlaybackInfo(
-        serverURL: URL,
-        itemId: String,
-        client: JellyfinClient,
-        body: PlaybackInfoDto
-    ) async throws -> PlaybackInfoResponse {
-        let url = serverURL.appendingPathComponent("/Items/\(itemId)/PlaybackInfo")
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        // Build the same auth header the SDK uses
-        var rawFields = [
-            "DeviceId": client.configuration.deviceID,
-            "Device": client.configuration.deviceName,
-            "Client": client.configuration.client,
-            "Version": client.configuration.version,
-        ]
-        if let token = client.accessToken {
-            rawFields["Token"] = token
-        }
-        let fields = rawFields.map { "\($0.key)=\($0.value)" }.joined(separator: ", ")
-        request.setValue("MediaBrowser \(fields)", forHTTPHeaderField: "Authorization")
-
-        // Encode the body
-        let encoder = JSONEncoder()
-        request.httpBody = try encoder.encode(body)
-
-        #if DEBUG
-        if let bodyStr = String(data: request.httpBody!, encoding: .utf8) {
-            let truncated = bodyStr.prefix(500)
-            debugLog("POST body: \(truncated)...")
-        }
-        #endif
-
-        let (data, urlResponse) = try await URLSession.shared.data(for: request)
-        let httpResponse = urlResponse as? HTTPURLResponse
-
-        #if DEBUG
-        debugLog("POST PlaybackInfo status: \(httpResponse?.statusCode ?? -1)")
-        if let responseStr = String(data: data, encoding: .utf8) {
-            debugLog("POST PlaybackInfo response body: \(responseStr.prefix(1000))")
-        }
-        #endif
-
-        guard let statusCode = httpResponse?.statusCode, (200..<300).contains(statusCode) else {
-            let statusCode = httpResponse?.statusCode ?? -1
-            #if DEBUG
-            if let bodyStr = String(data: data, encoding: .utf8) {
-                debugLog("PlaybackInfo \(statusCode) body: \(bodyStr.prefix(500))")
-            }
-            #endif
-            throw JellyfinError.playbackFailed("PlaybackInfo returned \(statusCode)")
-        }
-
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .custom { decoder in
-            let container = try decoder.singleValueContainer()
-            let dateStr = try container.decode(String.self)
-            let isoFormatter = ISO8601DateFormatter()
-            isoFormatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-            return isoFormatter.date(from: dateStr) ?? Date()
-        }
-        return try decoder.decode(PlaybackInfoResponse.self, from: data)
-    }
-
-    /// Fallback: direct stream URL without PlaybackInfo session.
-    private func buildDirectStreamURL(
-        itemId: String, serverURL: URL, token: String?, etag: String?
-    ) -> PlaybackInfo {
-        var components = URLComponents(url: serverURL, resolvingAgainstBaseURL: false) ?? URLComponents()
-        components.path = "/Videos/\(itemId)/stream"
-        var queryItems = [
-            URLQueryItem(name: "static", value: "true"),
-            URLQueryItem(name: "mediaSourceId", value: itemId),
-            URLQueryItem(name: "deviceId", value: deviceID),
-        ]
-        if let etag { queryItems.append(URLQueryItem(name: "tag", value: etag)) }
-        // Token passed via Authorization header on AVURLAsset, not in URL query params
-        components.queryItems = queryItems
-
-        let url = components.url ?? serverURL
-        #if DEBUG
-        debugLog("Direct stream fallback URL: \(url)")
-        #endif
-        return PlaybackInfo(url: url, playSessionId: nil, mediaSourceId: itemId, playMethod: .directStream,
-                            audioTracks: [], subtitleTracks: [], selectedAudioIndex: nil, selectedSubtitleIndex: nil,
-                            authToken: token)
-    }
-
-    // MARK: - Playback Reporting
-
-    public func reportPlaybackStart(itemId: String, userId: String, mediaSourceId: String?, playSessionId: String?, positionTicks: Int?, playMethod: PlayMethod) async {
-        guard let client = getClient() else { return }
-        let jellyfinMethod = JellyfinAPI.PlayMethod(rawValue: playMethod.rawValue)
-        let body = PlaybackStartInfo(
-            canSeek: true,
-            itemID: itemId,
-            mediaSourceID: mediaSourceId,
-            playMethod: jellyfinMethod,
-            playSessionID: playSessionId,
-            positionTicks: positionTicks
-        )
-        _ = try? await client.send(Paths.reportPlaybackStart(body))
-    }
-
-    public func reportPlaybackProgress(itemId: String, userId: String, mediaSourceId: String?, playSessionId: String?, positionTicks: Int?, isPaused: Bool, playMethod: PlayMethod) async {
-        guard let client = getClient() else { return }
-        let jellyfinMethod = JellyfinAPI.PlayMethod(rawValue: playMethod.rawValue)
-        let body = PlaybackProgressInfo(
-            canSeek: true,
-            isPaused: isPaused,
-            itemID: itemId,
-            mediaSourceID: mediaSourceId,
-            playMethod: jellyfinMethod,
-            playSessionID: playSessionId,
-            positionTicks: positionTicks
-        )
-        _ = try? await client.send(Paths.reportPlaybackProgress(body))
-    }
-
-    public func reportPlaybackStopped(itemId: String, userId: String, mediaSourceId: String?, playSessionId: String?, positionTicks: Int?) async {
-        guard let client = getClient() else { return }
-        let body = PlaybackStopInfo(
-            itemID: itemId,
-            mediaSourceID: mediaSourceId,
-            playSessionID: playSessionId,
-            positionTicks: positionTicks
-        )
-        _ = try? await client.send(Paths.reportPlaybackStopped(body))
-    }
-
-    /// Extracts audio and subtitle track info from a media source's stream list.
-    private static func extractTracks(from source: MediaSourceInfo) -> (audio: [MediaTrackInfo], subtitles: [MediaTrackInfo]) {
-        let streams = source.mediaStreams ?? []
-        let audio: [MediaTrackInfo] = streams
-            .filter { $0.type == .audio }
-            .compactMap { s in
-                guard let idx = s.index else { return nil }
-                let label = s.displayTitle ?? s.language ?? "Track \(idx)"
-                return MediaTrackInfo(id: idx, label: label, isDefault: s.isDefault ?? false, isForced: s.isForced ?? false)
-            }
-        let subtitles: [MediaTrackInfo] = streams
-            .filter { $0.type == .subtitle }
-            .compactMap { s in
-                guard let idx = s.index else { return nil }
-                let label = s.displayTitle ?? s.language ?? "Track \(idx)"
-                return MediaTrackInfo(id: idx, label: label, isDefault: s.isDefault ?? false, isForced: s.isForced ?? false)
-            }
-        return (audio, subtitles)
-    }
-
-    // Cached constant arrays — built once at class load time, reused on every playback request.
-    nonisolated(unsafe) private static let _directPlayProfiles: [DirectPlayProfile] = [
-        DirectPlayProfile(
-            audioCodec: "aac,ac3,alac,eac3,flac",
-            container: "mp4,m4v",
-            type: .video,
-            videoCodec: "h264,hevc"
-        ),
-        DirectPlayProfile(
-            audioCodec: "aac,ac3,alac,eac3,mp3,pcm_s16be,pcm_s16le,pcm_s24be,pcm_s24le",
-            container: "mov",
-            type: .video,
-            videoCodec: "h264,hevc,mjpeg"
-        ),
-        DirectPlayProfile(
-            audioCodec: "aac,ac3,eac3,mp3",
-            container: "mpegts",
-            type: .video,
-            videoCodec: "h264,hevc"
-        ),
-    ]
-    // hevc,h264 only — MPEG-4 ASP is not a valid HLS transcode target on Apple devices
-    // and causes Jellyfin to inject mpeg4-* URL parameters that AVFoundation doesn't recognise.
-    // enableSubtitlesInManifest = true → Jellyfin includes WebVTT renditions in the HLS manifest.
-    // AVKit shows them natively in ONE unified Subtitles menu on both iOS and tvOS.
-    // On iOS, HLSManifestLoader strips ASS/SSA tags from WebVTT segments.
-    // On tvOS, AVAssetResourceLoaderDelegate doesn't work, so ASS tags may appear in subtitles.
-    nonisolated(unsafe) private static let _transcodingProfiles: [TranscodingProfile] = [
-        TranscodingProfile(
-            audioCodec: "aac,ac3,alac,eac3,flac",
-            isBreakOnNonKeyFrames: true,
-            container: "mp4",
-            context: .streaming,
-            enableSubtitlesInManifest: true,
-            maxAudioChannels: "8",
-            minSegments: 2,
-            protocol: .hls,
-            type: .video,
-            videoCodec: "hevc,h264"
-        ),
-    ]
-    nonisolated(unsafe) private static let _subtitleProfiles: [SubtitleProfile] = [
-        SubtitleProfile(format: "srt",    method: .hls),
-        SubtitleProfile(format: "subrip", method: .hls),
-        SubtitleProfile(format: "vtt",    method: .hls),
-        SubtitleProfile(format: "webvtt", method: .hls),
-        SubtitleProfile(format: "ass",    method: .hls),
-        SubtitleProfile(format: "ssa",    method: .hls),
-        SubtitleProfile(format: "ttml",   method: .hls),
-        SubtitleProfile(format: "pgs",    method: .encode),
-        SubtitleProfile(format: "pgssub", method: .encode),
-        SubtitleProfile(format: "dvbsub", method: .encode),
-        SubtitleProfile(format: "dvdsub", method: .encode),
-        SubtitleProfile(format: "sub",    method: .encode),
-    ]
-
-    /// Builds a DeviceProfile matching Swiftfin's native player profile.
-    private static func buildAppleDeviceProfile(maxBitrate: Int = 40_000_000) -> DeviceProfile {
-        DeviceProfile(
-            directPlayProfiles: _directPlayProfiles,
-            maxStreamingBitrate: maxBitrate,
-            subtitleProfiles: _subtitleProfiles,
-            transcodingProfiles: _transcodingProfiles
-        )
-    }
-
     /// Invalidates every cached response (resume items, latest media, genres, etc.).
     /// Called by Settings → Server → Refresh Catalogue so that the next fetch hits
     /// the server rather than returning stale cached data.
@@ -1087,7 +208,7 @@ public final class JellyfinAPIClient: Sendable {
         #endif
     }
 
-    private var deviceID: String {
+    internal var deviceID: String {
         KeychainService.getOrCreateDeviceID()
     }
 
@@ -1095,4 +216,3 @@ public final class JellyfinAPIClient: Sendable {
         Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? "1.0.0"
     }
 }
-
