+++ Shared/Screens/LibraryHeroSection.swift
@@ -30,7 +30,7 @@ struct LibraryHeroSection: View {
             .overlay {
                 if let id = item.id {
                     CinemaLazyImage(
-                        url: appState.imageBuilder.imageURL(itemId: id, imageType: .backdrop, maxWidth: ImageURLBuilder.screenPixelWidth),
+                        url: appState.imageBuilder.imageURL(itemId: id, imageType: .backdrop, maxWidth: ImageURLBuilder.backdropPixelWidth),
                         fallbackIcon: nil,
                         fallbackBackground: CinemaColor.surfaceContainerLow
                     )
