+++ Shared/Screens/NativeVideoPresenter.swift
@@ -47,6 +47,8 @@ final class NativeVideoPresenter {
     private var playbackReporter: PlaybackReporter!
     private var skipSegments: SkipSegmentController!
     private var sleepTimer: SleepTimerController!
+    private var chapters: ChapterController!
+    private var endOfSeries: EndOfSeriesOverlayController!
 
     // Track state
     private var audioTracks: [MediaTrackInfo] = []
@@ -59,9 +61,9 @@ final class NativeVideoPresenter {
     // and PlaybackReporter.onTick from startProgressReporting.
     private var timeObserver: Any?
 
-    // End-of-series
+    // End-of-series — `currentSeriesName` is written by ChapterController after
+    // it fetches the full item; read by the itemEnd observer when auto-play ends.
     private var currentSeriesName: String?
-    private var finishedSeriesOverlay: UIView?
 
     // Debug: jump to last 15 seconds button
     private var skipToEndButton: UIButton?
@@ -119,6 +121,14 @@ final class NativeVideoPresenter {
             loc: loc,
             playerVCProvider: { [weak self] in self?.playerVC },
             onStopPlayback: { [weak self] in self?.playerVC?.dismiss(animated: true) }
+        )
+        self.chapters = ChapterController(
+            apiClient: apiClient, userId: userId, imageBuilder: imageBuilder
+        )
+        self.endOfSeries = EndOfSeriesOverlayController(
+            loc: loc,
+            playerVCProvider: { [weak self] in self?.playerVC },
+            onDone: { [weak self] in self?.playerVC?.dismiss(animated: true) }
         )
     }
 
@@ -234,7 +244,12 @@ final class NativeVideoPresenter {
             startProgressReporting()
             observeItemEnd(playerItem, player: avPlayer)
             self.skipSegments.load(for: self.itemId)
-            fetchAndApplyChapters(for: self.itemId, playerItem: playerItem)
+            self.chapters.fetchAndApply(
+                itemId: self.itemId,
+                playerItem: playerItem,
+                token: self.playbackInfo?.authToken,
+                onSeriesNameResolved: { [weak self] name in self?.currentSeriesName = name }
+            )
             self.sleepTimer.startIfNeeded()
             showSkipToEndButtonIfDebugEnabled()
         }
@@ -388,7 +403,7 @@ final class NativeVideoPresenter {
         // Debug: jump to the last 15 seconds. Lives in the transport bar custom menu
         // (rather than as a free-floating UIView overlay) so the tvOS focus engine
         // can actually reach it via the Siri Remote.
-        if UserDefaults.standard.bool(forKey: "debug.showSkipToEnd") {
+        if UserDefaults.standard.bool(forKey: SettingsKey.debugShowSkipToEnd) {
             items.append(UIAction(
                 title: "⏭ Skip to End (Debug)",
                 image: UIImage(systemName: "forward.end.alt.fill")
@@ -508,7 +523,12 @@ final class NativeVideoPresenter {
             startProgressReporting()
             observeItemEnd(playerItem, player: avPlayer)
             skipSegments.load(for: ep.id)
-            fetchAndApplyChapters(for: ep.id, playerItem: playerItem)
+            chapters.fetchAndApply(
+                itemId: ep.id,
+                playerItem: playerItem,
+                token: self.playbackInfo?.authToken,
+                onSeriesNameResolved: { [weak self] name in self?.currentSeriesName = name }
+            )
             // Episode navigation restarts the sleep timer (keeps playback "session" alive).
             sleepTimer.startIfNeeded()
             showSkipToEndButtonIfDebugEnabled()
@@ -555,13 +575,13 @@ final class NativeVideoPresenter {
         ) { [weak self] _ in
             MainActor.assumeIsolated {
                 guard let self else { return }
-                let autoPlay = UserDefaults.standard.object(forKey: "autoPlayNextEpisode") as? Bool ?? true
+                let autoPlay = UserDefaults.standard.object(forKey: SettingsKey.autoPlayNextEpisode) as? Bool ?? SettingsKey.Default.autoPlayNextEpisode
                 if autoPlay, let next = self.nextEpisode, self.episodeNavigator != nil {
                     self.navigateToEpisode(next)
                 } else if autoPlay, self.episodeNavigator != nil, self.nextEpisode == nil,
                           let seriesName = self.currentSeriesName {
                     // We just finished the last episode of a series while auto-play is on.
-                    self.showFinishedSeriesOverlay(seriesName: seriesName)
+                    self.endOfSeries.show(seriesName: seriesName)
                 }
             }
         }
@@ -600,155 +620,6 @@ final class NativeVideoPresenter {
         #endif
     }
 
-    // Shared overlay styling (used by the iOS finishedSeriesOverlay).
-
-    private var overlayCardWidth: CGFloat {
-        #if os(tvOS)
-        640
-        #else
-        340
-        #endif
-    }
-
-    private var overlayIconSize: CGFloat {
-        #if os(tvOS)
-        72
-        #else
-        48
-        #endif
-    }
-
-    private var overlayTitleSize: CGFloat {
-        #if os(tvOS)
-        36
-        #else
-        22
-        #endif
-    }
-
-    private var overlaySubtitleSize: CGFloat {
-        #if os(tvOS)
-        22
-        #else
-        15
-        #endif
-    }
-
-    private var overlayButtonSize: CGFloat {
-        #if os(tvOS)
-        24
-        #else
-        16
-        #endif
-    }
-
-    // MARK: - Chapters
-
-    /// Fetches the full item (which carries the chapter list) and builds an
-    /// `AVNavigationMarkersGroup` exposed to AVPlayerViewController's scrubber.
-    /// Chapter images are downloaded in parallel and embedded as artwork metadata.
-    private func fetchAndApplyChapters(for id: String, playerItem: AVPlayerItem) {
-        let client = apiClient
-        let uid = userId
-        let builder = imageBuilder
-        let token = playbackInfo?.authToken
-        Task { [weak self, weak playerItem] in
-            guard let fullItem = try? await client.getItem(userId: uid, itemId: id) else { return }
-
-            // Capture the series name while we have the full item — used by the
-            // end-of-series completion overlay when this was the last episode.
-            await MainActor.run { [weak self] in
-                self?.currentSeriesName = fullItem.seriesName
-            }
-
-            guard let chapters = fullItem.chapters, chapters.count > 1 else { return }
-
-            // Chapter markers are tvOS-only (AVNavigationMarkersGroup lives in AVKit
-            // on tvOS only). Skip image download and marker build on iOS.
-            #if os(tvOS)
-            // Fetch chapter thumbnails in parallel. Missing images are fine — chapter
-            // markers still render with just a title.
-            let images: [Int: Data] = await withTaskGroup(of: (Int, Data?).self) { group in
-                for (index, _) in chapters.enumerated() {
-                    let url = builder.chapterImageURL(itemId: id, imageIndex: index, maxWidth: 480)
-                    group.addTask {
-                        await Self.loadChapterImage(url: url, token: token).map { (index, $0) } ?? (index, nil)
-                    }
-                }
-                var results: [Int: Data] = [:]
-                for await (idx, data) in group {
-                    if let data { results[idx] = data }
-                }
-                return results
-            }
-
-            await MainActor.run {
-                guard let self, let playerItem else { return }
-                self.applyChapterMarkers(chapters: chapters, images: images, to: playerItem)
-            }
-            #else
-            _ = builder
-            _ = token
-            _ = playerItem
-            _ = self
-            #endif
-        }
-    }
-
-    /// Downloads one chapter thumbnail, attaching the Jellyfin access token so the server
-    /// authorises the request. Returns `nil` on HTTP error or non-image content.
-    nonisolated private static func loadChapterImage(url: URL, token: String?) async -> Data? {
-        var request = URLRequest(url: url)
-        if let token {
-            request.addValue("MediaBrowser Token=\(token)", forHTTPHeaderField: "Authorization")
-        }
-        guard let (data, response) = try? await URLSession.shared.data(for: request),
-              let http = response as? HTTPURLResponse, http.statusCode == 200 else {
-            return nil
-        }
-        return data
-    }
-
-    /// Builds `AVTimedMetadataGroup` markers for the scrubber and assigns them via
-    /// `AVPlayerItem.navigationMarkerGroups`. Group title becomes the section label
-    /// in AVKit's Chapters menu.
-    ///
-    /// **tvOS-only**: `AVNavigationMarkersGroup` lives in `AVKit` and only ships on tvOS.
-    /// On iOS, `AVPlayerViewController` has no built-in chapters scrubber, so we skip.
-    private func applyChapterMarkers(chapters: [ChapterInfo], images: [Int: Data], to playerItem: AVPlayerItem) {
-        #if os(tvOS)
-        var markers: [AVTimedMetadataGroup] = []
-        markers.reserveCapacity(chapters.count)
-
-        for (index, chapter) in chapters.enumerated() {
-            let startSeconds = Double(chapter.startPositionTicks ?? 0) / 10_000_000
-            let startTime = CMTime(seconds: startSeconds, preferredTimescale: 600)
-            let range = CMTimeRange(start: startTime, duration: .zero)
-
-            var items: [AVMetadataItem] = []
-
-            let titleItem = AVMutableMetadataItem()
-            titleItem.identifier = .commonIdentifierTitle
-            titleItem.value = (chapter.name ?? "Chapter \(index + 1)") as NSString
-            titleItem.extendedLanguageTag = "und"
-            items.append(titleItem)
-
-            if let data = images[index] {
-                let artwork = AVMutableMetadataItem()
-                artwork.identifier = .commonIdentifierArtwork
-                artwork.value = data as NSData
-                artwork.dataType = kCMMetadataBaseDataType_JPEG as String
-                artwork.extendedLanguageTag = "und"
-                items.append(artwork)
-            }
-
-            markers.append(AVTimedMetadataGroup(items: items, timeRange: range))
-        }
-
-        let group = AVNavigationMarkersGroup(title: "Chapters", timedNavigationMarkers: markers)
-        playerItem.navigationMarkerGroups = [group]
-        #endif
-    }
 
     // MARK: - Error Recovery
 
@@ -804,7 +675,7 @@ final class NativeVideoPresenter {
         // tvOS path is handled by the transport-bar custom menu; nothing to draw here.
         return
         #else
-        let enabled = UserDefaults.standard.bool(forKey: "debug.showSkipToEnd")
+        let enabled = UserDefaults.standard.bool(forKey: SettingsKey.debugShowSkipToEnd)
         guard enabled else {
             hideSkipToEndButton()
             return
@@ -852,155 +723,6 @@ final class NativeVideoPresenter {
         guard duration.isFinite, duration > 15 else { return }
         let target = CMTime(seconds: duration - 15, preferredTimescale: 600)
         player.seek(to: target, toleranceBefore: .zero, toleranceAfter: .zero)
-    }
-
-    // MARK: - End-of-Series Overlay
-
-    /// Shown when the last episode of a series finishes with autoplay on. Gives the user
-    /// a concrete "you're done" moment rather than the player just sitting at the end.
-    private func showFinishedSeriesOverlay(seriesName: String) {
-        #if os(tvOS)
-        // Same focus-engine reasoning as `showSleepOverlay` — UIAlertController
-        // owns its own focus context so the Done button is reachable with the remote.
-        guard finishedSeriesOverlay == nil, let vc = playerVC else { return }
-
-        let alert = UIAlertController(
-            title: String(format: loc.localized("player.finishedSeries.title"), seriesName),
-            message: loc.localized("player.finishedSeries.subtitle"),
-            preferredStyle: .alert
-        )
-        alert.addAction(UIAlertAction(
-            title: loc.localized("player.finishedSeries.done"),
-            style: .default,
-            handler: { [weak self] _ in
-                self?.finishedSeriesOverlay = nil
-                self?.playerVC?.dismiss(animated: true)
-            }
-        ))
-
-        finishedSeriesOverlay = alert.view
-        vc.present(alert, animated: true)
-        return
-        #else
-        guard finishedSeriesOverlay == nil, let vc = playerVC else { return }
-
-        let container = UIView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.backgroundColor = UIColor.black.withAlphaComponent(0.75)
-        container.alpha = 0
-
-        let card = UIView()
-        card.translatesAutoresizingMaskIntoConstraints = false
-        card.layer.cornerRadius = 20
-        card.clipsToBounds = true
-
-        let blur = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
-        blur.translatesAutoresizingMaskIntoConstraints = false
-        blur.isUserInteractionEnabled = false
-        card.addSubview(blur)
-
-        let icon = UIImageView(image: UIImage(systemName: "checkmark.circle.fill"))
-        icon.tintColor = .systemGreen
-        icon.translatesAutoresizingMaskIntoConstraints = false
-        icon.contentMode = .scaleAspectFit
-
-        let titleLabel = UILabel()
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel.font = .systemFont(ofSize: overlayTitleSize, weight: .bold)
-        titleLabel.textColor = .white
-        titleLabel.textAlignment = .center
-        titleLabel.numberOfLines = 0
-        titleLabel.text = String(format: loc.localized("player.finishedSeries.title"), seriesName)
-
-        let subtitleLabel = UILabel()
-        subtitleLabel.translatesAutoresizingMaskIntoConstraints = false
-        subtitleLabel.font = .systemFont(ofSize: overlaySubtitleSize, weight: .medium)
-        subtitleLabel.textColor = .white.withAlphaComponent(0.8)
-        subtitleLabel.textAlignment = .center
-        subtitleLabel.numberOfLines = 0
-        subtitleLabel.text = loc.localized("player.finishedSeries.subtitle")
-
-        var doneConfig = UIButton.Configuration.plain()
-        var doneTitle = AttributedString(loc.localized("player.finishedSeries.done"))
-        doneTitle.font = .systemFont(ofSize: overlayButtonSize, weight: .semibold)
-        doneTitle.foregroundColor = UIColor.black
-        doneConfig.attributedTitle = doneTitle
-        doneConfig.contentInsets = NSDirectionalEdgeInsets(top: 14, leading: 24, bottom: 14, trailing: 24)
-        doneConfig.background.backgroundColor = UIColor.white.withAlphaComponent(0.95)
-        doneConfig.background.cornerRadius = 12
-
-        let doneButton = UIButton(
-            configuration: doneConfig,
-            primaryAction: UIAction { [weak self] _ in
-                self?.hideFinishedSeriesOverlay()
-                self?.playerVC?.dismiss(animated: true)
-            }
-        )
-        doneButton.translatesAutoresizingMaskIntoConstraints = false
-
-        card.addSubview(icon)
-        card.addSubview(titleLabel)
-        card.addSubview(subtitleLabel)
-        card.addSubview(doneButton)
-        container.addSubview(card)
-
-        let targetView = vc.view!
-        targetView.addSubview(container)
-
-        NSLayoutConstraint.activate([
-            container.topAnchor.constraint(equalTo: targetView.topAnchor),
-            container.bottomAnchor.constraint(equalTo: targetView.bottomAnchor),
-            container.leadingAnchor.constraint(equalTo: targetView.leadingAnchor),
-            container.trailingAnchor.constraint(equalTo: targetView.trailingAnchor),
-
-            card.centerXAnchor.constraint(equalTo: container.centerXAnchor),
-            card.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-            card.widthAnchor.constraint(equalToConstant: overlayCardWidth),
-
-            blur.topAnchor.constraint(equalTo: card.topAnchor),
-            blur.bottomAnchor.constraint(equalTo: card.bottomAnchor),
-            blur.leadingAnchor.constraint(equalTo: card.leadingAnchor),
-            blur.trailingAnchor.constraint(equalTo: card.trailingAnchor),
-
-            icon.topAnchor.constraint(equalTo: card.topAnchor, constant: 36),
-            icon.centerXAnchor.constraint(equalTo: card.centerXAnchor),
-            icon.widthAnchor.constraint(equalToConstant: overlayIconSize),
-            icon.heightAnchor.constraint(equalToConstant: overlayIconSize),
-
-            titleLabel.topAnchor.constraint(equalTo: icon.bottomAnchor, constant: 20),
-            titleLabel.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            titleLabel.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-
-            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 8),
-            subtitleLabel.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            subtitleLabel.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-
-            doneButton.topAnchor.constraint(equalTo: subtitleLabel.bottomAnchor, constant: 28),
-            doneButton.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            doneButton.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-            doneButton.bottomAnchor.constraint(equalTo: card.bottomAnchor, constant: -24),
-        ])
-
-        self.finishedSeriesOverlay = container
-
-        UIView.animate(withDuration: 0.3) { container.alpha = 1 }
-        #endif
-    }
-
-    private func hideFinishedSeriesOverlay() {
-        guard let container = finishedSeriesOverlay else { return }
-        #if os(tvOS)
-        if let presented = playerVC?.presentedViewController {
-            presented.dismiss(animated: true)
-        }
-        finishedSeriesOverlay = nil
-        return
-        #else
-        UIView.animate(withDuration: 0.25, animations: { container.alpha = 0 }) { _ in
-            container.removeFromSuperview()
-        }
-        finishedSeriesOverlay = nil
-        #endif
     }
 
     // MARK: - Helpers
@@ -1078,7 +800,8 @@ final class NativeVideoPresenter {
         skipSegments.teardown()
         hideSkipToEndButton()
         sleepTimer.teardown()
-        hideFinishedSeriesOverlay()
+        chapters.teardown()
+        endOfSeries.teardown()
         playerObservation?.invalidate()
         playerObservation = nil
         if let obs = itemEndObserver {
